requirements.txt
bleach==6.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-graphiql-debug-toolbar==0.2.0
django-mptt==0.14.0
django-pglocks==1.0.4
django-prometheus==2.3.1
django-redis==5.4.0
django-rich==1.8.0
django-rq==2.10.1
django-taggit==5.0.1
django-tables2==2.7.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
feedparser==6.0.11
graphene-django==3.0.0
gunicorn==21.2.0
Jinja2==3.1.3
Markdown==3.5.2
mkdocs-material==9.5.7
mkdocstrings[python-legacy]==0.24.0
netaddr==0.10.1
Pillow==10.2.0
psycopg[binary,pool]==3.1.18
PyYAML==6.0.1
requests==2.31.0
social-auth-app-django==5.4.0
social-auth-core[openidconnect]==4.5.2
svgwrite==1.4.3
tablib==3.5.0
tzdata==2023.4